requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.3.1
django-jalali==6.0.1
django-jalali-date==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jdatetime==4.1.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
